Vendor all go commands together
Having a go.mod file per cmd was responsible depending on this
repo within the go.mod file. This prevents immediately consuming
changes in the api or pkg directories and generally seemed odd.

Import paths are still target/strelka instead of sublime-security/strelka
so a replace is used in go.mod. We could change this. It should matter
whether the repos local directory is in target or sublime-security.