Only tag full version number
The docker files pull a lot of dependencies which make them
susceptable to transiently breaking. It'll be important that
we test images adequetly before "releasing". Since the major/minor
and major tags would update immediately this could cause a new
image to be consumed before any testing. If we need these tags
we should move them to the update latest workflow.

Of course we can also put more effort into making the build reliable,
but that's more involved.